fixing missed frees and invalid reads
