+++ Makefile
@@ -6,7 +6,7 @@
 #    By: jadithya <jadithya@student.42abudhabi.ae>  +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/07/18 15:46:23 by jadithya          #+#    #+#              #
-#    Updated: 2023/09/13 17:06:54 by jadithya         ###   ########.fr        #
+#    Updated: 2023/09/25 12:48:38 by jadithya         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -30,6 +30,7 @@ SRCS := $(SRCDIR)/minishell.c\
 		$(SRCDIR)/find_cmd.c\
 		$(SRCDIR)/execute_cmd.c\
 		$(SRCDIR)/redirs.c\
+		$(SRCDIR)/free.c\
 		$(SRCDIR)/$(BUILTINS)/pwd.c\
 		$(SRCDIR)/$(BUILTINS)/export.c\
 		$(SRCDIR)/$(BUILTINS)/cd.c\
