+++ src/builtins/echo.c
@@ -6,7 +6,7 @@
 /*   By: jadithya <jadithya@student.42abudhabi.ae>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/09 16:19:42 by jadithya          #+#    #+#             */
-/*   Updated: 2023/09/10 15:05:33 by jadithya         ###   ########.fr       */
+/*   Updated: 2023/09/25 16:58:11 by jadithya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
