+++ src/builtins/env.c
@@ -6,7 +6,7 @@
 /*   By: jadithya <jadithya@student.42abudhabi.ae>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/21 19:40:20 by jebucoy           #+#    #+#             */
-/*   Updated: 2023/09/17 17:23:03 by jadithya         ###   ########.fr       */
+/*   Updated: 2023/09/25 16:58:18 by jadithya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
