+++ src/builtins/exit.c
@@ -6,7 +6,7 @@
 /*   By: jadithya <jadithya@student.42abudhabi.ae>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/09 16:14:38 by jadithya          #+#    #+#             */
-/*   Updated: 2023/09/17 16:11:52 by jadithya         ###   ########.fr       */
+/*   Updated: 2023/09/25 16:58:25 by jadithya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
