+++ src/builtins/export.c
@@ -6,7 +6,7 @@
 /*   By: jadithya <jadithya@student.42abudhabi.ae>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/19 20:31:10 by jadithya          #+#    #+#             */
-/*   Updated: 2023/09/17 16:16:08 by jadithya         ###   ########.fr       */
+/*   Updated: 2023/09/25 16:58:31 by jadithya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
