+++ src/builtins/pwd.c
@@ -6,7 +6,7 @@
 /*   By: jadithya <jadithya@student.42abudhabi.ae>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/19 18:44:49 by jadithya          #+#    #+#             */
-/*   Updated: 2023/09/09 14:14:42 by jadithya         ###   ########.fr       */
+/*   Updated: 2023/09/25 16:57:49 by jadithya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
