+++ src/run_cmds.c
@@ -6,7 +6,7 @@
 /*   By: jadithya <jadithya@student.42abudhabi.ae>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/29 15:53:08 by jadithya          #+#    #+#             */
-/*   Updated: 2023/09/24 22:23:27 by jadithya         ###   ########.fr       */
+/*   Updated: 2023/09/25 13:02:02 by jadithya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
